Crate robonomics_primitives

Low-level types used throughout the Robonomics code.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them.

Amount

The amount type, should be signed version of balance.

AssetId

Id used for identifying assets.

Balance

Balance of an account.

Block

Block type.

BlockId

Block ID.

BlockNumber

An index to a block.

DigestItem

Digest item type.

Hash

A hash of some data used by the chain.

Header

Header type.

Index

Index of a transaction in the chain.

Moment

Type used for expressing timestamp.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

Timestamp

A timestamp: milliseconds since the unix epoch. u64 is enough to represent a duration of half a billion years, when the time scale is milliseconds.